Hide pages before user connects wallet
